# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FishTrackPro API",
        "description": "API для приложения FishTrackPro - социальной сети для рыболовов",
        "contact": {
            "name": "FishTrackPro Support",
            "url": "https://fishtrackpro.ru",
            "email": "support@fishtrackpro.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.fishtrackpro.ru/api/v1",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Получить список групп",
                "description": "Возвращает список доступных групп",
                "operationId": "45b53f3c45b7730f9dfc37b5a224df20",
                "responses": {
                    "200": {
                        "description": "Список групп",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Рыболовы Москвы"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Группа для московских рыболовов"
                                                    },
                                                    "cover_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/cover.jpg"
                                                    },
                                                    "privacy": {
                                                        "type": "string",
                                                        "example": "public"
                                                    },
                                                    "members_count": {
                                                        "type": "integer",
                                                        "example": 150
                                                    },
                                                    "owner": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Иван Иванов"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Создать новую группу",
                "description": "Создает новую группу рыболовов",
                "operationId": "2831b9923d7fbb3854e15c6094bacf52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 191,
                                        "example": "Рыболовы Москвы"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Группа для московских рыболовов"
                                    },
                                    "cover_url": {
                                        "type": "string",
                                        "maxLength": 512,
                                        "example": "https://example.com/cover.jpg"
                                    },
                                    "privacy": {
                                        "type": "string",
                                        "enum": [
                                            "public",
                                            "private",
                                            "closed"
                                        ],
                                        "example": "public"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Группа успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Group created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Рыболовы Москвы"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Группа для московских рыболовов"
                                                },
                                                "privacy": {
                                                    "type": "string",
                                                    "example": "public"
                                                },
                                                "owner_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/groups/{id}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Получить детали группы",
                "description": "Возвращает подробную информацию о конкретной группе",
                "operationId": "2ac653904738468ee509c67eab015ad7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID группы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали группы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Рыболовы Москвы"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Группа для московских рыболовов"
                                                },
                                                "cover_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/cover.jpg"
                                                },
                                                "privacy": {
                                                    "type": "string",
                                                    "example": "public"
                                                },
                                                "members_count": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "posts_count": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "owner": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Иван Иванов"
                                                        },
                                                        "username": {
                                                            "type": "string",
                                                            "example": "ivan_ivanov"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "members": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Петр Петров"
                                                            },
                                                            "role": {
                                                                "type": "string",
                                                                "example": "member"
                                                            },
                                                            "joined_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Группа не найдена"
                    }
                }
            }
        },
        "/groups/{id}/join": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Присоединиться к группе",
                "description": "Позволяет пользователю присоединиться к группе рыболовов",
                "operationId": "94a01cc2cacf4368b3f9ecca038ccc88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID группы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно присоединились к группе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Joined group successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка присоединения",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Already a member"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация"
                    },
                    "404": {
                        "description": "Группа не найдена"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/groups/{id}/leave": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Покинуть группу",
                "description": "Позволяет пользователю покинуть группу рыболовов",
                "operationId": "6c60a30ee10997ae6260d10a8aa4cd71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID группы",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно покинули группу",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Left group successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка покидания",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not a member of this group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется аутентификация"
                    },
                    "404": {
                        "description": "Группа не найдена"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "JWT Bearer token для аутентификации",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "jwt": []
        }
    ]
}